Updated types

+++ src/parsers/obj.cpp
@@ -11,7 +11,7 @@ using namespace uinta;
 
 namespace uinta {
 struct objface {
-  int vert, uv, norm, index;
+  i32 vert, uv, norm, index;
 
   objface() {
     vert = -1;
@@ -25,21 +25,21 @@ struct objface {
   }
 };
 
-void extractLineFloats(const std::vector<std::string>& lines, float* const buffer, const uint size, const char delimiter);
-int findOrInsertFaceData(const objface&, std::vector<objface>&, const uint);
-void packColors(const MeshAttrib&, float* const, uint* cont, const std::vector<objface>&, const float* const);
-void packNormals(const MeshAttrib&, float* const, uint* cont, const std::vector<objface>&, const float* const);
-void packUVs(const MeshAttrib&, float* const, uint* cont, const std::vector<objface>&, const float* const);
-void packVertices(const MeshAttrib&, float* const, uint* cont, const std::vector<objface>&, const float* const, bool);
+void extractLineFloats(const std::vector<std::string>& lines, f32* const buffer, const u32 size, const char delimiter);
+i32 findOrInsertFaceData(const objface&, std::vector<objface>&, const u32);
+void packColors(const MeshAttrib&, f32* const, u32* cont, const std::vector<objface>&, const f32* const);
+void packNormals(const MeshAttrib&, f32* const, u32* cont, const std::vector<objface>&, const f32* const);
+void packUVs(const MeshAttrib&, f32* const, u32* cont, const std::vector<objface>&, const f32* const);
+void packVertices(const MeshAttrib&, f32* const, u32* cont, const std::vector<objface>&, const f32* const, bool);
 void parseFile(const std::string&, std::vector<std::string>&, std::vector<std::string>&, std::vector<std::string>&,
                std::vector<std::string>&);
 void processFaceStrs(std::string*, std::vector<objface>&);
-void processFaces(const std::vector<std::string>&, std::vector<objface>&, uint* const, uint* const);
+void processFaces(const std::vector<std::string>&, std::vector<objface>&, u32* const, u32* const);
 
 }  // namespace uinta
 
-void uinta::loadObj(const std::string& objBuffer, float* const vbuf, uint* vcount, uint* const ibuf, uint* icount,
-                    uint* const ioff, const std::unordered_map<MeshAttribType, MeshAttrib>& attribs) {
+void uinta::loadObj(const std::string& objBuffer, f32* const vbuf, u32* vcount, u32* const ibuf, u32* icount, u32* const ioff,
+                    const std::unordered_map<MeshAttribType, MeshAttrib>& attribs) {
   if (!attribs.size()) {
     SPDLOG_WARN("Unable to parse .obj file: No attributes provided!");
     return;
@@ -56,13 +56,13 @@ void uinta::loadObj(const std::string& objBuffer, float* const vbuf, uint* vcoun
 
   bool hasColor = vertexLines.size() > 0 && 5 == std::count(vertexLines.at(0).begin(), vertexLines.at(0).end(), ' ');
 
-  float vertices[vertexLines.size() * (hasColor ? 6 : 3)];
+  f32 vertices[vertexLines.size() * (hasColor ? 6 : 3)];
   extractLineFloats(vertexLines, vertices, hasColor ? 6 : 3, ' ');
 
-  float uvs[uvLines.size() * 2];
+  f32 uvs[uvLines.size() * 2];
   extractLineFloats(uvLines, uvs, 2, ' ');
 
-  float normals[normalLines.size() * 3];
+  f32 normals[normalLines.size() * 3];
   extractLineFloats(normalLines, normals, 3, ' ');
 
   std::vector<objface> faceData;
@@ -86,9 +86,9 @@ void uinta::loadObj(const std::string& objBuffer, float* const vbuf, uint* vcoun
   }
 }
 
-void uinta::processFaces(const std::vector<std::string>& faceLines, std::vector<objface>& result, uint* const indexbuffer,
-                         uint* const ioff) {
-  for (uint i = 0, len = faceLines.size(); i < len; i++) {
+void uinta::processFaces(const std::vector<std::string>& faceLines, std::vector<objface>& result, u32* const indexbuffer,
+                         u32* const ioff) {
+  for (u32 i = 0, len = faceLines.size(); i < len; i++) {
     auto tmp = std::string(faceLines.at(i));
 
     std::string strs[3];
@@ -108,48 +108,48 @@ void uinta::processFaces(const std::vector<std::string>& faceLines, std::vector<
   *ioff += faceLines.size() * 3;
 }
 
-void uinta::packColors(const MeshAttrib& attrib, float* const vbuf, uint* const vcount, const std::vector<objface>& faceData,
-                       const float* const vertexData) {
-  for (int i = 0, len = faceData.size(); i < len; i++) {
+void uinta::packColors(const MeshAttrib& attrib, f32* const vbuf, u32* const vcount, const std::vector<objface>& faceData,
+                       const f32* const vertexData) {
+  for (i32 i = 0, len = faceData.size(); i < len; i++) {
     auto index = faceData.at(i).vert - 1;
     if (index < 0) break;
-    memcpy(&vbuf[attrib.offset + attrib.stride * i], &vertexData[index * 6 + 3], 3 * sizeof(float));
+    memcpy(&vbuf[attrib.offset + attrib.stride * i], &vertexData[index * 6 + 3], 3 * sizeof(f32));
     *vcount += 3;
   }
 }
 
-void uinta::packNormals(const MeshAttrib& attrib, float* const vbuf, uint* const vcount, const std::vector<objface>& faceData,
-                        const float* const normalData) {
-  for (int i = 0, len = faceData.size(); i < len; i++) {
+void uinta::packNormals(const MeshAttrib& attrib, f32* const vbuf, u32* const vcount, const std::vector<objface>& faceData,
+                        const f32* const normalData) {
+  for (i32 i = 0, len = faceData.size(); i < len; i++) {
     auto index = faceData.at(i).norm - 1;
     if (index < 0) break;
-    memcpy(&vbuf[attrib.offset + attrib.stride * i], &normalData[index * 3], 3 * sizeof(float));
+    memcpy(&vbuf[attrib.offset + attrib.stride * i], &normalData[index * 3], 3 * sizeof(f32));
     *vcount += 3;
   }
 }
 
-void uinta::packUVs(const MeshAttrib& attrib, float* const vbuf, uint* const vcount, const std::vector<objface>& faceData,
-                    const float* const uvData) {
-  for (int i = 0, len = faceData.size(); i < len; i++) {
+void uinta::packUVs(const MeshAttrib& attrib, f32* const vbuf, u32* const vcount, const std::vector<objface>& faceData,
+                    const f32* const uvData) {
+  for (i32 i = 0, len = faceData.size(); i < len; i++) {
     auto index = faceData.at(i).uv - 1;
     if (index < 0) break;
-    memcpy(&vbuf[attrib.offset + attrib.stride * i], &uvData[index * 3], 3 * sizeof(float));
+    memcpy(&vbuf[attrib.offset + attrib.stride * i], &uvData[index * 3], 3 * sizeof(f32));
     *vcount += 3;
   }
 }
 
-void uinta::packVertices(const MeshAttrib& attrib, float* const vbuf, uint* const vcount, const std::vector<objface>& faceData,
-                         const float* const vertexData, bool hasColor) {
-  for (int i = 0, len = faceData.size(); i < len; i++) {
+void uinta::packVertices(const MeshAttrib& attrib, f32* const vbuf, u32* const vcount, const std::vector<objface>& faceData,
+                         const f32* const vertexData, bool hasColor) {
+  for (i32 i = 0, len = faceData.size(); i < len; i++) {
     auto index = faceData.at(i).vert - 1;
     if (index < 0) break;
-    memcpy(&vbuf[attrib.offset + attrib.stride * i], &vertexData[index * (hasColor ? 6 : 3)], 3 * sizeof(float));
+    memcpy(&vbuf[attrib.offset + attrib.stride * i], &vertexData[index * (hasColor ? 6 : 3)], 3 * sizeof(f32));
     *vcount += 3;
   }
 }
 
-int uinta::findOrInsertFaceData(const objface& face, std::vector<objface>& faceData, const uint max) {
-  for (int i = 0; i < max; i++) {
+i32 uinta::findOrInsertFaceData(const objface& face, std::vector<objface>& faceData, const u32 max) {
+  for (u32 i = 0; i < max; i++) {
     if (faceData[i] == face) return i;
     if (faceData[i] == objface()) {
       faceData[i].vert = face.vert;
@@ -163,7 +163,7 @@ int uinta::findOrInsertFaceData(const objface& face, std::vector<objface>& faceD
 }
 
 void uinta::processFaceStrs(std::string* values, std::vector<objface>& result) {
-  for (int i = 0; i < 3; i++) {
+  for (i32 i = 0; i < 3; i++) {
     std::string tmp = std::string(values[i]);
     std::string vert = tmp.substr(0, tmp.find('/'));
     tmp.erase(0, tmp.find('/') + 1);
@@ -212,10 +212,10 @@ void uinta::parseFile(const std::string& objBuffer, std::vector<std::string>& ve
   }
 }
 
-void uinta::extractLineFloats(const std::vector<std::string>& lines, float* const buffer, const uint size, const char delimiter) {
-  for (int i = 0, len = lines.size(); i < len; i++) {
+void uinta::extractLineFloats(const std::vector<std::string>& lines, f32* const buffer, const u32 size, const char delimiter) {
+  for (size_t i = 0, len = lines.size(); i < len; i++) {
     auto tmp = std::string(lines.at(i));
-    for (int j = 0; j < size; j++) {
+    for (size_t j = 0; j < size; j++) {
       buffer[i * size + j] = std::stof(tmp.substr(0, tmp.find(delimiter)));
       tmp.erase(0, tmp.find(delimiter) + 1);
     }
